extract Track
